+++ src/y/elf/datafunctions/MaxFunction.java
@@ -10,13 +10,13 @@ public class MaxFunction extends DataFunction {
 	}
 	
 	@Override
-	public int function(int[] data) {
+	public double function(int[] data) {
 		int m = data[0];
 		
 		for (int i=1; i<data.length; i++)
 			if (data[i] > m)
 				m = data[i];
 		
-		return m;
+		return (double)m;
 	}
 }
