+++ src/y/elf/datafunctions/MinFunction.java
@@ -9,13 +9,13 @@ public class MinFunction extends DataFunction {
 	}
 
 	@Override
-	public int function(int[] data) {
+	public double function(int[] data) {
 		int m = data[0];
 		
 		for (int i=1; i<data.length; i++)
 			if (data[i] < m)
 				m = data[i];
 		
-		return m;
+		return (double)m;
 	}
 }
